requirements.txt
django-registration == 1.0
Unipath == 1.0
PIL == 1.1.7